requirements.txt
requests>=2.31.0
jsonschema>=4.21.1
google-genai
pytest>=8.0.0
yfinance>=0.2.36
pandas>=2.0.0
PyYAML>=6.0.1
markdown>=3.5.0
statsmodels
scikit-learn
matplotlib
